requirements.txt
click==6.7
dnspython==1.15.0
dominate==2.3.1
eventlet==0.24.1
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-SocketIO==3.0.1
greenlet==0.4.14
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
monotonic==1.5
Pillow==5.2.0
python-engineio==2.2.0
python-socketio==2.0.0
simplejson==3.16.0
six==1.11.0
visitor==0.1.3
Werkzeug==0.14.1
xlrd==1.1.0
